emotion-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
websockets
pypdf
google-generativeai